requirements.txt
fastapi
uvicorn
requests
python-dotenv
pyyaml
Pillow
streamlit